apps/operator_console/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.38
pydantic==2.10.6
pyjwt==2.10.1
requests==2.32.3
python-multipart==0.0.20

